Update modelfetch and add modelfetch-hono library
- Updated `modelfetch` library to include a CLI entry point.
- Renamed `modelfetch` to `modelfetch CLI` in documentation for clarity.
- Added new `modelfetch-hono` library with initial implementation and configuration.
- Updated example applications to include `modelfetch` as a devDependency.
- Adjusted TypeScript configuration to reference the new `modelfetch-hono` library.
- Simplified `modelfetch` source code to log a message for initial testing.

This commit enhances the modularity of the project by introducing a dedicated library for Hono integration.
