+++ apps/example-js/package.json
@@ -6,7 +6,9 @@
   "main": "./src/index.js",
   "dependencies": {
     "@modelcontextprotocol/sdk": "^1.13.2",
-    "modelfetch": "workspace:^",
     "zod": "^3.25.67"
+  },
+  "devDependencies": {
+    "modelfetch": "workspace:^"
   }
 }
