+++ libs/modelfetch-hono/eslint.config.js
@@ -1,0 +1,5 @@
+import createESLintConfig from "create-eslint-config";
+
+const eslintConfig = createESLintConfig();
+
+export default eslintConfig;
