+++ libs/modelfetch-hono/tsconfig.lib.json
@@ -1,0 +1,10 @@
+{
+  "include": ["src"],
+  "references": [],
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "sourceRoot": "/libs/modelfetch-hono/src",
+    "rootDir": "src",
+    "outDir": "dist"
+  }
+}
