+++ libs/modelfetch/bin/index.js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+await import("../dist/index.js");
